+++ esp-hal/src/soc/esp32/peripherals.rs
@@ -65,6 +65,7 @@ crate::peripherals! {
     SYSTEM <= DPORT,
     TIMG0 <= TIMG0,
     TIMG1 <= TIMG1,
+    TOUCH <= virtual,
     TWAI0 <= TWAI0,
     UART0 <= UART0,
     UART1 <= UART1,